heres a happy medium to allow the code to still be rendered but just not ran

+++ jobs/yb-tserver/templates/bin/pre_start.erb.sh
@@ -2,10 +2,11 @@
 
 set -eu
 
-<% if p("rm_post_install_completed_on_startup") %>
-echo "pre_start will be removing the following .post_install.sh.completed markers before re-initializing post_install.sh..."
-find -L /var/vcap/packages /var/vcap/data/packages -type f -name ".post_install.sh.completed" -print -delete
-<% end %>
+RM_POST_INSTALL_COMPLETED_ON_STARTUP=<%= p("rm_post_install_completed_on_startup") %>
+if [ "${RM_POST_INSTALL_COMPLETED_ON_STARTUP}" = "true" ]; then
+  echo "pre_start will be removing the following .post_install.sh.completed markers before re-initializing post_install.sh..."
+  find -L /var/vcap/packages /var/vcap/data/packages -type f -name ".post_install.sh.completed" -print -delete
+fi
 
 /var/vcap/packages/yugabyte/bin/post_install.sh
 /var/vcap/jobs/yb-tserver/bin/cert-linker.sh
